Update syllabifier.py
The Vi_rule focuses on word-initial vowel syllables. So I used the index 0 to indicate the first letter of the word.

The Vii_rule is focused only on vowels. So, I indicated that the last letter before the syllable boundary should be a vowel that is preceded by another vowel. Thus, I used -1 and -2 as vowels. I hope it makes sense.